refact: Change setMinMaxValue parameters from double to int
Updated the setMinMaxValue method to accept integer parameters instead of doubles. Also adjusted the labels and slider values to use these integer parameters accordingly.

+++ randomizer-desktop/randomizer/src/main/java/com/revortix/randomizer/ui/view/component/MinMaxSlider.java
@@ -50,7 +50,7 @@ public class MinMaxSlider extends HBox {
         rangeSlider.setMin(value);
     }
 
-    public void setMinMaxValue(double min, double max) {
+    public void setMinMaxValue(int min, int max) {
         minLabel.setText(min + "s");
         maxLabel.setText(max + "s");
         rangeSlider.setLowValue(min);
